+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/trace/ReflectionProcessor.java
@@ -257,7 +257,7 @@ class ReflectionProcessor extends AbstractProcessor {
                 break;
             }
 
-            case "putBundleInCache":
+            case "loadBundleOf":
             case "getBundleImpl": {
                 expectSize(args, 5);
                 String baseName = (String) args.get(2);